+++ docs/CHANGELOG.rst
@@ -44,7 +44,8 @@ Added
 - Add FastQC option ``nogroup`` to ``bbduk-single`` and ``bbduk-paired`` processes
 - Add CPM normalization to ``htseq-count-raw`` process
 - Add ``workflow-bbduk-star-htseq-paired``
-- Add ``prepare-annotation-dexseq`` process
+- Add ``prepare-annotation-dexseq`` and
+  ``differentialexpression-dexseq`` processes
 
 Fixed
 -----